+++ harvest/experts-client/lib/query/grant_role/bind.rq
@@ -21,7 +21,7 @@ select distinct (?relationship as ?RELATIONSHIP__) ?filename ?type where {
           .
   }
 
-  bind(replace(str(?relationship),str(oap:),"ark:/87287/d7mh2m/grantee/") as ?id)
+  bind(replace(str(?relationship),str(oap:),"ark:/87287/d7mh2m/relationship/") as ?id)
   filter(regex(?username,"@ucdavis.edu"))
   bind(concat('expert/',md5(?username),"/",?id,".jsonld.json") as ?filename)
 }
