+++ services/base-service/models/grant_role/index.js
@@ -1,7 +1,7 @@
-const GranteeModel = require('./model.js');
+const GrantRoleModel = require('./model.js');
 module.exports = {
   api : require('./api.js'),
-  model : new GranteeModel(),
+  model : new GrantRoleModel(),
   schema : require('../base/schema/minimal.json'),
   transform: require('../base/transform.js')
 }
